requirements.txt
asgiref==3.8.1
base==0.0.0
bcrypt==4.1.2
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
PyJWT==2.8.0
python-slugify==8.0.4
sqlparse==0.5.0
text-unidecode==1.3
django-filter
dj-rest-auth